Hawaiian amakihi point transect data

Also known as the Common 'Amakihi, a type of Hawaiian honeycreeper

Format

A data.frame with 1487 rows and 12 variables

Region.Label strata names (seven strata)

Area size of study area (set to 0)

Sample.Label transect ID

Effort number of visits to point

object object ID

distance radial distance (m)

Month month survey conducted (not used)

OBs observer ID (note capitalisation of variable name)

Sp species code (COAM) for all detections

MAS Time after sunrise (min)

HAS Time after sunrise (hours)

Study.Area name of study area

Note

Example for investigating covariates in the detection function. Note high colinearity between two measures of time since sunrise. Convergence problems can result from models with several factor covariates.

References

Marques, T.A., L. Thomas, S.G. Fancy and S.T. Buckland. (2007) Improving estimates of bird density using multiple-covariate distance sampling. The Auk 124 (4): 1229–1243. doi:10.1642/0004-8038(2007)124[1229:IEOBDU]2.0.CO;2