Update cookies preferences

Create bins from a set of binned distances and a set of cutpoints.

This is an internal routine and shouldn't be necessary in normal analyses.

Usage

create_bins(data, cutpoints)

Arguments

data

data.frame with at least the column distance.

cutpoints

vector of cutpoints for the bins

Value

argument data with two extra columns distbegin and distend.

Author

David L. Miller

Examples

if (FALSE) { # \dontrun{
library(Distance)
data(minke)

# put the minke data into bins 0-1, 1-2, 2-3 km
minke_cuts <- create_bins(minke[!is.na(minke$distance),], c(0,1,2,3))
} # }