Multiplier bootstrap helper functions

Helper to use a models specified using activity::fitact to fit an activity model and generate single realisations for bootstrapping with bootdht.

Usage

make_activity_fn(..., detector_daily_duration = 24)

Arguments

...

parameters specified by activity::fitact

detector_daily_duration

by default we assume that detectors were able to detect animals for 24 hours, if they were only able to do this for some proportion of the day (say daylight hours), then adjust this argument accordingly

Value

a function which generates a single bootstrap estimate of availability

Details

Uses activity::fitact to generate single possible availability estimates based on bootstraps. The function returns another function, which can be passed to bootdht. It is recommended that you try out the function before passing it to bootdht. See examples for a template for use.

Author

David L Miller