Print summary of distance detection function model object

Provides a brief summary of a distance sampling analysis. Including: detection function parameters, model selection criterion, and optionally abundance in the covered (sampled) region and its standard error.

Usage

# S3 method for class 'summary.dsmodel'
print(x, ...)

Arguments

x

a summary of distance sampling analysis

...

unspecified and unused arguments for S3 consistency

Value

Nothing, just prints the summary.

See also

Author

David L. Miller and Jeff Laake